Fix failed CI
