+++ app/models/kaui/killbill_registerable.rb
@@ -1,17 +1,8 @@
 # frozen_string_literal: true
 
-# rubocop:disable Style/ClassAndModuleChildren, Style/OneClassPerFile
+require 'devise/models/killbill_registerable'
 
-# Hack for Zeitwerk - must define Kaui::KillbillRegisterable before Devise::Models::KillbillRegisterable
+# Placeholder for Zeitwerk - must define Kaui::KillbillRegisterable before Devise::Models::KillbillRegisterable
 module Kaui
   module KillbillRegisterable; end
 end
-
-module Devise
-  module Models
-    module KillbillRegisterable
-      include Registerable
-    end
-  end
-end
-# rubocop:enable Style/ClassAndModuleChildren, Style/OneClassPerFile
