+++ lib/devise/models/killbill_registerable.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Devise
+  module Models
+    module KillbillRegisterable
+      include Registerable
+    end
+  end
+end
